Refactor parser and lexer into separate modules
This commit refactors the `parser.rs` file by extracting the lexer
and related components into a new `lexer.rs` file. This separation
improves modularity and resolves a circular dependency between the
parser and the AST.

The tests for both the parser and the lexer have been moved into their
own respective files, `parser_tests.rs` and `lexer_test.rs`, and are
now included as test modules.

The necessary import paths in `main.rs` and `ast.rs` have been
updated to reflect the new file structure.
